linkBio/requirements.txt
reflex==0.5.7
python-dotenv
supabase